refactor(mcp): consume SDK static-endpoint failover
Replace the private failoverClient with client.NewLanternFailover from
sdks/go (#592). The comma-separated LANTERN_ADDR contract stays in mcp/
via parseLanternAddrs; the failover policy now lives in the SDK.

Closes #593
